builders.sh: Run individual build.sh scripts with -x flags
builders.sh itself has "set -ex", but as soon as we call into per-image
build scripts, useful tracing stops. Instead of patching all the individual
scripts, just run them with "bash -x" (as the name "build.sh" suggests, it
should be a shell script, so interpreting it with bash shouldn't be a
regression).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I0ed79bc42f1931a6d3db454156540c52b15505ea