ci-dockerslaves-cleanup: add MISRA update from ECR
As part of the docker image cleanup routine on the Jenkins agents,
we should be attempting to update any of the MISRA images from
the private ECR so that they're available for Jenkins use.

This changes adds in an update script in misra-docker-images/update.sh,
and then adds it to the periodic job run to clean up all Jenkins agents.

Note that this change applies to production rather than staging even
though the MISRA images are only used in staging.  This is because both
production and staging use the same Jenkins agents, and the cleanup script
is only run from production to prevent competing cleanup processes from
interfering w/ each other.

Signed-off-by: Kelley Spoon <kelley.spoon@linaro.org>
Change-Id: Icbad25863176413d30bc0289884d0ad037c20bcd