ci-dockerslaves-cleanup: backout broken misra update
The misra images update script is currently not working as
the AWS credentials need to be added to enable authentication.

Since that work is going to take longer than an hour, let's
backout the call to the update script so we can avoid
job failure alerts being delivered in triplicate.

Signed-off-by: Kelley Spoon <kelley.spoon@linaro.org>
Change-Id: Iddda2d660383ca95cce95d5f5d19c7e7331d5d35