ci-dockerfiles-deployment: Remove to recreate as a different job type
This job is currently a matrix job with a single configuration, which makes
its usage and maintenance a bit confusing. We can't just change the type of
existing job from matrix to normal ("freestyle") in Jenkins, so will remove
and then immediately recreate as such.

This is similar to the ci-dockerfiles-sanity-check change done previously.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I102e92ba90aadab36b9ca57dadce2291464c2158